Use the dates from the query string
We were using the min and max dates returned from
content-data-api to determine the length of the x axis
for the time-series charts.

This causes an error if the time-series for a metric is
empty or missing.

Also the steer from designer is that it would be better for the
chart to show the entire date range even if the data does not
fill that time range.

This commit refactors the MetricsController and ChartPresenter
to use the dates specified in the URL.