Fix bug metrics formatting bug
When satisfaction score is not recieved due to existing problems with that metric
then currently we have an error raised in the format_metric_value method.

This PR ensures that this error will not be raised and that our app can handle nil
values for metrics which are measured as percentages.