Change link text as per designs
Remove the word `visit` from the link to content item on gov.uk link on
the single content item page.

The word `visit` is necessary for accessibility purposes but not helpful for
the design so this PR hides it using the visually-hidden class to allow
screen readers to understand it.