Use symbols in API response
In our codebase / tests we are not being consistent in using symbols
or string in hashes. Sometimes we even change the code to accommodate
symbols in our Fixtures.

This is not a very good practice, so I have symbolised the response
in origin (right after it is returned by the Adaptor) so that we can
be consistent in our codebase using symbols.