Improve the "Make changes to the page" functionality
Make the URL functional, at least for Whitehall Publisher, and
Publisher content. For other publishing apps, it won't appear, and
should be extended when possible.

To help with isolating the logic, I've moved it in to a helper. This
makes the choice regarding what URL and label to provide to the view.