Fix search rate bug
Search rates were being calculated incorrectly due to our use of the method
`number_with_delimiter` which, for example, returns '10,000' when passed 10000.

Calling `to_f` on '10,000' returns 10 and this is the cause of our current incorrect
calculation of search_rate on some pages.

In this PR this is solved by formatting the `total_searches` and `total_pviews` into
a float before using them to calculate `on_page_search_rate`.

This PR adds a test for this functionality and makes changes to default data in order
to test that this bug is not longer a problem.