Use the table helper for the content page
Change the content index page to use the TableHelper class from the
govuk_publishing_components gem. This should help add more
functionality to the table, sorting for example, and make it display
better.

This is a work in progress, the styling for the header is a bit off,
and sorting doesn't work yet.