Rework params handling in the ContentController
Take a more rigerous approach to processing the parameters, permitting
the ones that are expected, and merging them with a set of defaults.

The parameters are passed through to the FindContent class, so this
makes what is happening in the controller clearer. I think this also
fixes a bug where the default for `date_range` wasn't being passed
through to the FindContent class.

This change is motivated by adding another default for
organisation_id.