Fix searches_per_visit_bug
We currently have a bug which raises an error when calculating searches per page
if we do not have any data for either pviews or upviews for a certain period.

This bug has raised it's head when we set the date range to a period longer than we
have data for. This is because we will not have any data for the previous period, which
we use to calculate trends and so we get the error which is raised by this bug:

'`NoMethodError: undefined method `zero?' for nil:NilClass'

This PR calls `.to_f` on the value field of pviews and upviews. This will convert nil to 0
which will stop the error being raised if we receive a nil value.