Return ‘no comparison data’ if no comparison data
Add logic to single_content_item_presenter to return ‘no comparison data’
if there is no data from the previous period from which current data can be compared.

Returning a special value if we have no data to compare to will allow for changes in
the front end logic to handle this edge case in following commits.