Handle ‘no comparison data’ within views
Handle the edge case where we have no previous data with which to
compare current data with within the UI.

As discussed with our designers we will be displaying `no comparison data’
string in place of the trend percentage figure in the glance metrics and will not
display anything at all in place of the trend percentage and time period on the
timeseries charts.

Add tests and test data to assert the logic is working as expected across the
various partials.