Handle incomplete previous data
Stop rendering a trend percentage if the data from the previous
period is not complete.

In order to show the direction in which each metric is trending we
need to have access to data from the previous period. The reason
for this change is that  if this `previous` data is incomplete it is likely
to render a misleading trend result.