Fix last-month comparison data bug
At a glance trends currentl returns `no comparison data` when the
date period `last-month` is selected. This is because we checked for
complete data based on the number of days in each month being the
same. Obviously this is not the case.

This PR changes the logic for the `past-month` date range in order to
check that the previous month has data for each day in the month,
regardless of the length of the month and regardless of how many
days are in the current month.