Fix satisfaction responses bug
We were not showing the number of responses to user satisfaction surveys
in our UI. This PR allows us to show the number of responses.

The problem was caused by us referencing `satisfaction_responses` when
in the api response the correct key is `satisfaction_score_responses`.

This PR corrects this error and updates the tests to reflect this.