Show full selected date range in charts
We want to show the selected date range in the
metrics charts.

The correct way to do this would be to get the x axis
of the chart to recognise that it is rendering dates,
then to give a a min and max values for the x axis.

I've tried the first part of this and it works fine.
However the ChartKick js has a bug in it.

When passing the min and max values to google charts,
it mangles the date and the google chart ends up with an
empty object.

We have a work around which is to pad out the data with dummy
values for all the dates in the date range.

This also affects the table.