Fix 404 page
Initial symptom of this issue was that we were not showing the page title in the browser
tab when we search for the metric of a page that does not exist in our data warehouse.

After some investigation I found we did not have the correct static html pages and were
not rendering them correctly. We were rendering them with our application layout applied.

This PR adds a static 404 page (the same one as is used by content publisher) and
ensures we render it without our application layout. This addresses the issue of not having
a title in the browser tab too.