Handle filter description for all/no orgs.
CSV Download was failing when `All organisations` or
`No primary organisation` is selected.

This was because it was trying to get the organisation
name to generate the filename.

`all` and `none` do not exist in the list of organisations so this was
failing.

Short-circuit for `all` and `none` to return `All organisations` and
`No organisation`.