Move index page feature tests
The feature tests for the index page is very long, so the last test
added in the previous commit is split out alongside the small chunk of
related tests. Have also moved index page feature tests into an index
folder, so that we can then split the file out further.

This not a full refactor but just a change to aid review
for now. We have imminent work to do a full sweep of refactorings.