Clarify feedback comments short title
User research showed that users expect 'About feedback comments' to contain the comments themselves. Changing the short title to 'Number of feedback comments' will help clarify that it's information about the metric. I've checked the other places this is used (index page, at a glance metrics) and we have space for the longer title.