Fix handling all document types for the CSV download
I'm looking at this, as it led to a subtle issue where the CSV files
would be empty, if you filtered by all document types. FindContent is
used for both the HTML pages, and the CSV download, but this
conditional being removed only applied to the HTML page data.

To keep the Content Data API in the Content Performance Manager
consistent, this has been extended [1] to support 'all' as a special
value for document_type, similar to how 'all' is handled for
organisation_id.

With that change, the issue should be resolved, but also move the
special case for document_type in the Content Data Admin for
completeness and to make the code paths more consistent.

1: https://github.com/alphagov/content-performance-manager/commit/477bfea231de72892bae3fd213ae66acc93ca6b2