Fix nil context error when no satisfaction metrics
When we have no values for `satisfaction_yes` or `satisfaction_no` we should not
fail, so I am converting those values to 0. This is consistent with how the user
understands the metric.

I have not called `to_i` on the method `value_for` because not all the metrics
should be converted to 0.