Abbreviate glance metric figures
Glance metrics were written to take input of an abbreviated number
and the appropriate labels, but we have never passed in the data in
that format. This means that the text is often stretching the limits
of its container and occasionally overflows. This change fixes that,
using a presenter to send the correct data to the glance metric
component where it is necessary to abbreviate.