Abbreviate glance metric figures
Glance metrics were written to take input of an abbreviated number
and the appropriate labels, but we have never passed in the data in
that format. This means that the text is often stretching the limits
of its container and occasionally overflows. This change lets the
component work out and display the abbreviated numbers itself.