Send a counter to StatsD for csv download time
We want to know how long the users are waiting
for a CSV download.

This commit adds `start_time` to the Sidekiq worker call.
The Sidekiq worker then uses this to send a counter to StatsD
with the number of seconds elapsed.
It sends a second counter when the elapsed time is over 60 seconds.