Support `local_env.yml`
Local (mac not VM) development is difficult.

This commit adds support for `local_env.rb` (like CPM) for
both test and development environments.

The file is git-ignored so we can all have different ones.

This enables you to have local environment variables.

You don't have to use it but it makes life easier if you do.