Improve layout of page data page which has no data to show
When a content is published there is no data present until the ETL process
executes. Currently, the page does not hint that there is no data present,
this may confuse users to think that the page maybe broken.

Now the page is styled like other pages for consistency and displays the sentence,
no data, to make it more clear.