Add sassc-rails as an explicit dependency
This app was relying on this dependency through
govuk_publishing_components which seemed unusual as it has a direct
need. This makes the dependency more explicit. It's also worth noting
the sassc-rails is no longer a dependency of newer versions of
govuk_publishing_components since [1]

As is conventional across various GOV.UK apps I've added sassc-rails as
one of the framework dependencies in dependabot.

[1]: https://github.com/alphagov/govuk_publishing_components/pull/1479