Include yarn linting in default Rake task
This runs a yarn install on Jenkins so that JS and Sass linting can run
there. It also disables sassLint since this no longer needs to be run.

I updated the Rakefile to be a bit more conventional with GOV.UK apps,
by doing lint as the first action and not using Rails.env for
disable the default task.