WIP: specify a target-branch to allow security PRs back through
According to [the docs](https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/configuration-options-for-the-dependabot.yml-file#target-branch),
Dependabot will ignore the configs in dependabot.yml for security
PRs, if a target-branch has been set to something other than the
default branch.

I'm hoping to accomplish two things with these changes:

1) Allow security PRs through again (even if they are for
   dependencies not on the `allow` list)
2) Have a way of distinguishing between security and non security
   related PRs (by the absence of the new `not-security` label).

Apart from the verbosity of the config, this does cause one new
issue: what to do with the `latest-dependencies` branch (we'll
definitely have to automate merging that to `main`, or at least
automate opening a PR to `main`). We'll also need to figure out
what happens if the `latest-dependencies` branch has not been
created yet.