Comment out GovukContentSecurityPolicy
GOV.UK hadn't intended for this app to have the GOV.UK Content Security
Policy yet, with us first planning to roll out this to frontend app. It
looks like this was added as part of an outsourced Rails update [1],
where the dev couldn't have known about our nuanced context.

As this is an app that doesn't receive a lot of developer attention I'm
disabling this as I don't want breaking changes to the CSP [2] to end up
in this app.

[1]: https://github.com/alphagov/content-data-admin/pull/812
[2]: https://github.com/alphagov/govuk_app_config/pull/279