Remove GA tracking from accessible-autocomplete
This code was never actually used in this app:

The `track` function is only called if `data-track-category` and `data-track-action` are both set on the <select> element. These attributes would be set in `app/views/components/_accessible_autocomplete.html.erb` if a `data_attributes` local is passed in. I've had a look through all the views where we include this partial, and none of them pass any locals when rendering the partial.

Given this, I'm going to remove the tracking code, which also allows us to remove the entire `onConfirm` function (as the remainder of the function is a workaround for a bug caused by the fact that we override `onConfirm`).