Add a service to get all policy information
Policies are bespoke content and accessibility issues created for
GOV.UK. These policies are created and maintained by GDS content
designers in Siteimprove.

When the policy issues for a page are returned by the API, it doesn't
return the description of the policy, just the policy ID. Though the
policy description is show in the html page report.

The only way to get an individual policy description is get them all
from the API and filter them by policd ID.

This service will be needed to create and object that has mapped the
policy violation to its description so that it can be displayed on the
page.

There are currently over 450 policies defined for GOV.UK content. The
API response paginates the response. The page size has been set to 500
to get around this, however this will need to be solved properly if the
number policies gets much bigger.

Co-authored-by: Keith Lawrence <keith.lawrence@digital.cabinet-office.gov.uk>
Co-authored-by: Leena Gupte <leena.gupte@digital.cabinet-office.gov.uk>
Co-authored-by: Max Harden <max.harden@digital.cabinet-office.gov.uk>