Add a summary presenter for Siteimprove
Adds a summary presenter to facilitate displaying links to the
Siteimprove page reports and to enable linking to a specific issue in
the page report.

The link to the accessibility report is not correct in the Siteimprove
API. The API incorrectly returns:

https://my2.siteimprove.com/Inspector/1054012/Accessibility/Page?pageId=2&impmd=0

as the link rather than:

https://my2.siteimprove.com/Inspector/1054012/A11y/Page?pageId=2&impmd=0&conf=a+aa+aria+si

so we need to modify it.

The other links are returned correctly.

Co-authored-by: Keith Lawrence <keith.lawrence@digital.cabinet-office.gov.uk>
Co-authored-by: Leena Gupte <leena.gupte@digital.cabinet-office.gov.uk>
Co-authored-by: Max Harden <max.harden@digital.cabinet-office.gov.uk>