Add Policy issues presenter
This presenter makes the policy information more readable.
For example, each policy has a name that consists of a code and short
name. These have been split so that only the short name will be displayed.

The full policy description is being pulled in from the Policies
service so that it can be easily displayed.

The policy code is also being used to split the policies in to style
and accessibility issues. This will allow the policy violations to be
rendered in different tables.

Co-authored-by: Keith Lawrence <keith.lawrence@digital.cabinet-office.gov.uk>
Co-authored-by: Leena Gupte <leena.gupte@digital.cabinet-office.gov.uk>
Co-authored-by: Max Harden <max.harden@digital.cabinet-office.gov.uk>