Add Quality Assurance Presenter
The direct link to the specific quality assurance issues are not
returned correctly in the summary API, therefore these direct links are
being stitched together here.

The `any?` method has been overloaded, because whilst the quality
assurance API returns many issues, were are only interested in
misspellings and broken links. If both of these are nil we should not
show the quality assurance section on the page at all, regardless of
any other quality assurance issues.

Co-authored-by: Keith Lawrence <keith.lawrence@digital.cabinet-office.gov.uk>
Co-authored-by: Leena Gupte <leena.gupte@digital.cabinet-office.gov.uk>
Co-authored-by: Max Harden <max.harden@digital.cabinet-office.gov.uk>