Add SiteimproveRedirectController
This controller will allow the GOV.UK browser extension [1] to link
directly to the page report.

At the moment the browser extension links to the Siteimprove search
page [2] as it has no way of finding the Siteimprove page id for a page.

The SiteimproveRedirectController inherits directly from the Base
Controller.

That is because the Application Controller does more things
than this controller needs.

The most important of those is that the Application Controller enforces
login via Signon. That's not appropriate here as you don't need a
Siteimprove account to view a page report, therefore you shouldn't need
a Signon account just to be redirected to Siteimprove website.

Rubocop doesn't like this behaviour, but it was thought to be better to
disable this cop than to have to remove the Signon check from the
Application Controller and specifically add it to all of the other
controllers.

[1]: https://github.com/alphagov/govuk-browser-extension
[2]: https://github.com/alphagov/govuk-browser-extension/pull/173

Co-authored-by: Keith Lawrence <keith.lawrence@digital.cabinet-office.gov.uk>
Co-authored-by: Leena Gupte <leena.gupte@digital.cabinet-office.gov.uk>
Co-authored-by: Max Harden <max.harden@digital.cabinet-office.gov.uk>