Remove GA data notice.
This removes the notice from views in the application and the email templates.
The notice refers to changes from 20 December 2019, which is over 4 years ago.
We only provide data from up to past 2 years so this notice is no longer
relevant.

We are also going to add a new notice around changes to data due to migration
from Universal Analytics to GA4/BigQuery.  Removing the old notice first before
re-introducting with new information may make users notice it more.