Add GA4 migration notice
The new Google GA4/BigQuery product we are migrating to are not direct
mappings of the old (Universal Analytics) metrics and the numbers may
vary.
Adding this notice should hopefully make it clear to the user of the
changes in case they see a significant change in pattern in the GOV.UK
pages they are monitoring.