Send CSV export durations to Prometheus
We have had some reports recently via Zendesk that users have not
received CSV downloads within 30 mins. We currently don't have an
overview on the performance of these export as StatsD has been
deprecated.

The CsvExportWorker has been updated so that it sends the CSV export
duration to Prometheus Pushgateway in the form of a histogram. We
can then visualise this on a Grafana dashboard so that it is easy
to monitor.

NB: We may need to update the bucket sizes once we have collected
a substantial amount of data.