Add distinct label to CSV export metrics
Prometheus scrapes data from Pushgateway every 60 seconds which
means it continues to record the same metric until another one is
sent. Adding a distinct label will allow us to filter out duplicate
values in Grafana.

As you can only register a histogram once, we have to create a new
one which allows us to add a label as part of it's setup.